API/requirements.txt
protobuf<3.20
numpy==1.21.6
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Keras-Preprocessing==1.1.2
tensorflow==2.8.2
Pillow==9.1.1
